fix tools output plus mobile layout
