+++ components/chat/chat-helpers/index.ts
@@ -10,7 +10,7 @@ import {
   buildFinalMessages,
   buildGoogleGeminiFinalMessages
 } from "@/lib/build-prompt"
-import { consumeReadableStream } from "@/lib/consume-stream"
+import { consumeReadableStream, parseDataStream } from "@/lib/consume-stream"
 import { Tables, TablesInsert } from "@/supabase/types"
 import {
   ChatFile,
@@ -313,46 +313,6 @@ export const fetchChatResponse = async (
   return response
 }
 
-function parseDataStream(line: string): { text: string; data: any } {
-  // regex to parse message like this '0: "text", 1: "text"'
-
-  const firstSeparatorIndex = line.indexOf(":")
-
-  if (firstSeparatorIndex === -1) {
-    throw new Error("Failed to parse stream string. No separator found.")
-  }
-
-  const prefix = line.slice(0, firstSeparatorIndex)
-
-  const streamPartsByCode = {
-    "0": {
-      parse: (jsonValue: JSONValue) => {
-        return { text: jsonValue as string, data: null }
-      }
-    },
-    "8": {
-      parse: (jsonValue: JSONValue) => {
-        return { data: jsonValue as any, text: "" }
-      }
-    }
-  }
-
-  if (
-    !Object.keys(streamPartsByCode).includes(
-      prefix as keyof typeof streamPartsByCode
-    )
-  ) {
-    throw new Error(`Failed to parse stream string. Invalid code ${prefix}.`)
-  }
-
-  const code = prefix as keyof typeof streamPartsByCode
-
-  const textValue = line.slice(firstSeparatorIndex + 1)
-  const jsonValue: JSONValue = JSON.parse(textValue)
-
-  return streamPartsByCode[code].parse(jsonValue)
-}
-
 export const processResponse = async (
   response: Response,
   lastChatMessage: ChatMessage,
@@ -376,6 +336,8 @@ export const processResponse = async (
         setFirstTokenReceived(true)
         setToolInUse("none")
 
+        console.log("chunk", chunk, selectedTools.length)
+
         try {
           contentToAdd = isHosted
             ? chunk
@@ -391,15 +353,25 @@ export const processResponse = async (
                   ""
                 )
 
+          console.log("contentToAdd", contentToAdd)
+
           if (contentToAdd === "") {
             return
           }
+
+          console.log(
+            "contentToAdd !== ''",
+            contentToAdd,
+            selectedTools.length > 0
+          )
 
           if (selectedTools.length > 0) {
             chunks.push(contentToAdd)
             if (chunk[chunk.length - 1] !== "\n") {
               return
             }
+
+            console.log("chunks", chunks)
 
             const streamParts = chunks
               .join("")
